Move things out of arch/dreamcast
