+++ include/kos/fs_random.h
