+++ kernel/arch/dreamcast/kernel/initall_hdrs.h
@@ -3,6 +3,7 @@
 #include <kos/fs.h>
 #include <kos/thread.h>
 #include <kos/fs_pty.h>
+#include <kos/fs_random.h>
 #include <kos/fs_romdisk.h>
 #include <kos/fs_ramdisk.h>
 #include <kos/library.h>
